Add initial project files
This commit includes the initial files for the PainXML project, which provides functionality for generating and parsing XML documents conforming to the ISO 20022 standard for payment initiation. Order and Document Go structs as well as associated conversion methods have been implemented. In addition, a README document for project overview and a LICENSE document for legal protections are included.